requirements.txt
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
libusb1==3.1.0
MarkupSafe==2.1.5
ndeflib==0.3.3
nfcpy==1.0.4
pyDes==2.0.1
pyserial==3.5
redis==5.0.6
requests==2.32.3
spotipy==2.24.0
urllib3==2.2.2
Werkzeug==3.0.3
